Migrate examples
