+++ examples/05_data_assimilation/01_stormcast_sda.py
@@ -6,7 +6,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,27 +16,28 @@
 
 # %%
 """
-StormCast Score-Based Data Assimilation
-=======================================
+StormCast-CONUS Score-Based Data Assimilation
+=============================================
 
-Running StormCast with guided diffusion posterior sampling to assimilate observations.
+Running StormCast-CONUS with guided diffusion posterior sampling to assimilate observations.
 
-This example demonstrates how to use the StormCast score-based data assimilation (SDA)
-model for convection-allowing regional forecasts that incorporate sparse in-situ
-observations using diffusion posterior sampling (DPS).
-Two forecasts are run—one without observations and one with ISD surface station data
-from central United States region to illustrate the impact of data assimilation.
+This example demonstrates how to use the StormCast-CONUS generative model with
+score-based data assimilation (SDA) over the Continental United States. Sparse
+in-situ surface observations from NOAA ISD are assimilated at each forecast step
+using diffusion posterior sampling (DPS) guidance.
+Two forecasts are run—one without observations and one with ISD surface station
+data from the central United States—to illustrate the impact of data assimilation.
 
 In this example you will learn:
 
-- How to load and initialise the StormCast score-based data assimilation (SDA) model
+- How to load StormCast-CONUS and configure SDA parameters
 - Fetching HRRR initial conditions and ISD surface observations
 - Running the model iteratively with and without observation assimilation
 - Comparing assimilated and non-assimilated forecasts
 """
 # /// script
 # dependencies = [
-#   "earth2studio[da-stormcast] @ git+https://github.com/NVIDIA/earth2studio.git",
+#   "earth2studio[data,stormcast-conus] @ git+https://github.com/NVIDIA/earth2studio.git",
 #   "cartopy",
 # ]
 # ///
@@ -46,15 +47,15 @@ In this example you will learn:
 # ------
 # This example requires the following components:
 #
-# - Assimilation Model: StormCast SDA :py:class:`earth2studio.models.da.StormCastSDA`.
+# - Prognostic Model: StormCast-CONUS :py:class:`earth2studio.models.px.StormCastCONUS`
+#   configured with SDA parameters.
 # - Datasource (state): HRRR analysis :py:class:`earth2studio.data.HRRR`.
 # - Datasource (obs): NOAA ISD surface stations :py:class:`earth2studio.data.ISD`.
-# - Datasource (conditioning): GFS forecasts :py:class:`earth2studio.data.GFS_FX`
-#   (loaded automatically by the model).
 #
-# StormCast score-based data assimilation (SDA) extends StormCast with diffusion
-# posterior sampling (DPS) guidance,
-# allowing sparse point observations to steer the generative diffusion process.
+# StormCast-CONUS extends the StormCast generative architecture to the full CONUS
+# HRRR domain at 3 km resolution. When an observation DataFrame is passed to each
+# generator step, diffusion posterior sampling (DPS) steers the denoising trajectory
+# toward the observed values.
 
 # %%
 import os
@@ -64,12 +65,10 @@ from dotenv import load_dotenv
 
 load_dotenv()  # TODO: make common example prep function
 
-from collections import OrderedDict
 from datetime import timedelta
 
 import numpy as np
 import torch
-import xarray as xr
 from loguru import logger
 from tqdm import tqdm
 
@@ -77,14 +76,30 @@ logger.remove()
 logger.add(lambda msg: tqdm.write(msg, end=""), colorize=True)
 
 from earth2studio.data import HRRR, ISD, fetch_data
-from earth2studio.models.da import StormCastSDA
-from earth2studio.utils.coords import map_coords_xr
+from earth2studio.models.px import StormCastCONUS
+from earth2studio.utils.coords import map_coords
 
-package = StormCastSDA.load_default_package()
-# Load the model onto the GPU and configure SDA
-# sda_std_obs: assumed (normalized) observation noise std (lower = trust obs more)
-# sda_gamma: DPS guidance scaling factor (lower = stronger assimilation)
-model = StormCastSDA.load_model(package, sda_std_obs=0.1, sda_gamma=1e-4)
+# Load the default model package
+package = StormCastCONUS.load_default_package()
+
+# Configure SDA: sda_std_obs is the assumed normalised observation noise std per
+# variable (lower = trust observations more).
+# sda_gamma is the DPS step-size scaling (lower = stronger guidance from obs).
+#
+# By default the example runs on the central-US subdomain to reduce GPU memory
+# requirements. For the full CONUS domain, comment out hrrr_lat_lim and
+# hrrr_lon_lim below (requires ~200 GB VRAM for SDA).
+hrrr_lat_lim = (273, 785)
+hrrr_lon_lim = (579, 1219)
+model = StormCastCONUS.load_model(
+    package,
+    hrrr_lat_lim=hrrr_lat_lim,  # comment out for full CONUS domain
+    hrrr_lon_lim=hrrr_lon_lim,  # comment out for full CONUS domain
+    num_diffusion_steps=18,
+    num_sda_diffusion_steps=96,
+    sda_std_obs=0.15,
+    sda_gamma=1e-3,
+)
 model = model.to("cuda:0")
 
 hrrr = HRRR()
@@ -93,87 +108,84 @@ hrrr = HRRR()
 # Fetch Initial Conditions
 # ------------------------
 # Pull HRRR analysis data for April 3rd 2025, a date that saw a major tornado
-# outbreak across the central United States, and select the sub-grid that
-# StormCast expects. The model's :py:meth:`init_coords` describes the required
+# outbreak across the central United States, and align it to the model's
 # coordinate system.
 
 # %%
 init_time = np.array([np.datetime64("2025-04-03T18:00")])
-ic = model.init_coords()[0]
 
-x = fetch_data(
+x, coords = fetch_data(
     hrrr,
     time=init_time,
-    variable=ic["variable"],
+    variable=model.variables,
     lead_time=np.array([np.timedelta64(0, "h")]),
     device="cuda:0",
-    legacy=False,
 )
-x = map_coords_xr(x, ic)
+x, coords = map_coords(x, coords, model.input_coords())
 
 # %%
 # Run Without Observations
 # ------------------------
-# Step the model forward 6 hours without any observations. This is equivalent to using
-# the StormCast prognostic model as it will just use EDM diffusion sampling under the
-# hood.  Each call to ``model.send(None)`` advances the state by one hour.
-# We store only the surface variables used for comparison (u10m, v10m, t2m).
+# Step the model forward 6 hours without any observations. This uses the
+# standard EDM diffusion sampler, equivalent to running StormCast-CONUS as a
+# pure generative forecast model. We store only the 10-m zonal wind (``u10m``)
+# used for comparison.
 
 # %%
 nsteps = 6
+plot_var = "u10m"
 plot_vars = ["u10m", "v10m", "t2m"]
+var_idx = list(model.variables).index(plot_var)
 
 np.random.seed(42)
 torch.manual_seed(42)
 if torch.cuda.is_available():
     torch.cuda.manual_seed_all(42)
 
-no_obs_frames = []
-gen = model.create_generator(x.copy())
-x_state = next(gen)  # Prime the generator, yields initial state
+no_obs_fields = []
+gen = model.create_generator(x.clone(), coords.copy())
+x_cur, c_cur = next(gen)  # prime the generator, yields initial state (lead_time = 0 h)
 
 for step in tqdm(range(nsteps), desc="No-obs forecast"):
-    logger.info(f"Running no-obs forecast step {step}")
-    x_state = gen.send(None)  # Advance one hour without observations
-    no_obs_frames.append(x_state.sel(variable=plot_vars).copy())
+    logger.info(f"Running no-obs forecast step {step + 1}/{nsteps}")
+    x_cur, c_cur = gen.send(None)  # advance one hour without observations
+    no_obs_fields.append(x_cur[0, 0, var_idx].cpu().numpy())
 
 gen.close()
-no_obs_da = xr.concat(no_obs_frames, dim="lead_time")
-
-# Save to Zarr (convert to numpy for storage)
-no_obs_np = no_obs_da.copy(data=no_obs_da.data.get())
-no_obs_np.to_dataset(name="prediction").to_zarr("outputs/21_no_obs.zarr", mode="w")
+no_obs_fields = np.stack(no_obs_fields)  # (nsteps, H, W)
 
 # %%
-# Fetch Observations and Run With Assimilation
+# Fetch Observations and Plot Station Locations
 # ---------------------------------------------
-# Fetch NOAA Integrated Surface Database (ISD) surface observations from the central
-# United States and assimilate them at each forecast step. The observations are fetched
-# for the valid time (initialisation time + lead time) so the model assimilates
-# temporally relevant data.
+# Fetch NOAA Integrated Surface Database (ISD) surface observations covering
+# the model domain.  The bounding box is derived from the model's lat/lon grid
+# so it automatically adjusts when running on a subdomain.
+# We visualise the station network before running the assimilation forecast.
 
 # %%
-stations = ISD.get_stations_bbox((32.0, -105.0, 45.0, -90.0))
+# ISD uses [-180, 180] longitude; model.lon is in [0, 360] — convert here.
+lat_min = float(model.lat.min())
+lat_max = float(model.lat.max())
+lon_min = float(model.lon.min()) - 360.0
+lon_max = float(model.lon.max()) - 360.0
+stations = ISD.get_stations_bbox((lat_min, lon_min, lat_max, lon_max))
 isd = ISD(stations=stations, time_tolerance=timedelta(minutes=15), verbose=False)
-
-# %%
-# Plot ISD Station Locations
-# --------------------------
-# Visualise the ISD stations that will provide observations for assimilation.
 
 # %%
 import cartopy
 import cartopy.crs as ccrs
 import matplotlib.pyplot as plt
 
-# Fetch a sample to get station locations
-sample_df = isd(init_time, ["t2m", "u10m", "v10m"])
+# Fetch a sample observation at the initial time to retrieve station positions
+sample_df = isd(init_time, plot_vars)
 station_lats = sample_df["lat"].values
 station_lons = sample_df["lon"].values
 
 plt.close("all")
 fig, ax = plt.subplots(subplot_kw={"projection": ccrs.PlateCarree()}, figsize=(8, 6))
-ax.set_extent([-110, -85, 30, 47], crs=ccrs.PlateCarree())
+ax.set_extent(
+    [lon_min - 2, lon_max + 2, lat_min - 2, lat_max + 2], crs=ccrs.PlateCarree()
+)
 ax.add_feature(
     cartopy.feature.STATES.with_scale("50m"), linewidth=0.5, edgecolor="black"
 )
@@ -187,16 +199,16 @@ ax.scatter(
     transform=ccrs.PlateCarree(),
     zorder=3,
 )
-ax.set_title("ISD Station Locations - Central United States")
-plt.savefig("outputs/21_isd_stations.jpg", dpi=150, bbox_inches="tight")
+ax.set_title("ISD Station Locations - StormCast-CONUS Domain")
+plt.savefig("outputs/03_isd_stations.jpg", dpi=150, bbox_inches="tight")
 
 # %%
 # Run Inference With Streaming Observations
 # ------------------------------------------
-# At each forecast step, fetch NOAA ISD observations for the current valid time
-# and send them to the model generator. The observations will be used in the SDA
-# guidance term when sampling the diffusion model effectively steering the generated
-# result to align with the stations data from the ISD data based.
+# At each forecast step the next valid time is determined from the current
+# generator state, ISD observations are fetched for that time, and the
+# observation DataFrame is sent to the generator. Zero-value wind reports
+# (anemometer failure) are filtered out before assimilation.
 
 # %%
 np.random.seed(42)
@@ -204,47 +216,44 @@ torch.manual_seed(42)
 if torch.cuda.is_available():
     torch.cuda.manual_seed_all(42)
 
-obs_frames = []
-gen = model.create_generator(x)
-x_state = next(gen)  # Prime the generator, yields initial state
+obs_fields = []
+gen = model.create_generator(x.clone(), coords.copy())
+x_cur, c_cur = next(gen)  # prime the generator, yields initial state
 
 for step in tqdm(range(nsteps), desc="Obs forecast"):
-    # Fetch observations for the next forecast step's valid time using model coords
-    x_coords = OrderedDict({d: x_state.coords[d].values for d in x_state.dims})
-    oc = model.output_coords((x_coords,))[0]
-    valid_time = oc["time"] + oc["lead_time"]
+    # Target valid time is one step ahead of the current generator state
+    valid_time = np.array(
+        [c_cur["time"][0] + c_cur["lead_time"][0] + np.timedelta64(1, "h")]
+    )
     obs_df = isd(valid_time, plot_vars)
-    logger.info(f"Running obs forecast step {step}, {len(obs_df)} obs")
-    x_state = gen.send(obs_df)  # Advance one hour with observations
-    obs_frames.append(x_state.sel(variable=plot_vars).copy())
+
+    # Drop zero-value wind reports (common anemometer failure mode)
+    if len(obs_df) > 0:
+        is_wind = obs_df["variable"].str.startswith(("u", "v"))
+        zero_wind = is_wind & (obs_df["observation"].abs() < 1e-5)
+        obs_df = obs_df[~zero_wind]
+
+    obs = obs_df if len(obs_df) > 0 else None
+    logger.info(
+        f"Step {step + 1}/{nsteps} — {len(obs_df) if obs is not None else 0} obs"
+    )
+    x_cur, c_cur = gen.send(obs)  # advance one hour with observations
+    obs_fields.append(x_cur[0, 0, var_idx].cpu().numpy())
 
 gen.close()
-obs_da = xr.concat(obs_frames, dim="lead_time")
-
-# Save to Zarr
-obs_np = obs_da.copy(data=obs_da.data.get())
-obs_np.to_dataset(name="prediction").to_zarr("outputs/21_with_obs.zarr", mode="w")
+obs_fields = np.stack(obs_fields)  # (nsteps, H, W)
 
 # %%
 # Post Processing
 # ---------------
-# Compare the two forecasts.  The top row shows the baseline forecast (no
-# observations), the middle row shows the assimilated forecast with observation
-# station locations overlaid as unfilled circles, and the bottom row shows the
-# difference (assimilated minus baseline).
+# Compare the two forecasts. The top row shows the baseline (no observations),
+# the middle row shows the assimilation forecast with station locations overlaid,
+# and the bottom row shows the signed difference (assimilated − baseline).
 
 # %%
 plt.close("all")
-variable = "u10m"
 
-# Load saved forecasts from Zarr stores
-no_obs_ds = xr.open_zarr("outputs/21_no_obs.zarr")
-obs_ds = xr.open_zarr("outputs/21_with_obs.zarr")
-
-no_obs_vals = no_obs_ds["prediction"].sel(variable=variable).values
-obs_vals = obs_ds["prediction"].sel(variable=variable).values
-
-# Lambert Conformal projection matching HRRR
+# HRRR Lambert Conformal projection
 projection = ccrs.LambertConformal(
     central_longitude=262.5,
     central_latitude=38.5,
@@ -260,14 +269,13 @@ fig, axes = plt.subplots(
 )
 fig.subplots_adjust(wspace=0.02, hspace=0.08, left=0.1, right=0.9)
 
+vmin, vmax = -10, 10
 for step in range(nsteps):
     lead_hr = step + 1
-    no_obs_field = no_obs_vals[0, step]
-    obs_field = obs_vals[0, step]
+    no_obs_field = no_obs_fields[step]
+    obs_field = obs_fields[step]
     diff_field = obs_field - no_obs_field
 
-    vmin = -10
-    vmax = 10
     # Row 0: No-obs forecast
     ax = axes[0, step]
     im0 = ax.pcolormesh(
@@ -307,7 +315,6 @@ for step in range(nsteps):
         linewidths=0.8,
         transform=ccrs.PlateCarree(),
         zorder=3,
-        label="Stations",
     )
     ax.add_feature(
         cartopy.feature.STATES.with_scale("50m"),
@@ -316,7 +323,7 @@ for step in range(nsteps):
         zorder=2,
     )
 
-    # Row 2: Difference (assimilated - baseline)
+    # Row 2: Difference (assimilated − baseline)
     ax = axes[2, step]
     im2 = ax.pcolormesh(
         model.lon,
@@ -334,79 +341,56 @@ for step in range(nsteps):
         zorder=2,
     )
 
-axes[0, 0].text(
-    -0.07,
-    0.5,
-    "No Obs",
-    va="bottom",
-    ha="center",
-    rotation="vertical",
-    rotation_mode="anchor",
-    fontsize=12,
-    transform=axes[0, 0].transAxes,
-)
-axes[1, 0].text(
-    -0.07,
-    0.5,
-    "Obs",
-    va="bottom",
-    ha="center",
-    rotation="vertical",
-    rotation_mode="anchor",
-    fontsize=12,
-    transform=axes[1, 0].transAxes,
-)
-axes[2, 0].text(
-    -0.07,
-    0.5,
-    "Difference",
-    va="bottom",
-    ha="center",
-    rotation="vertical",
-    rotation_mode="anchor",
-    fontsize=12,
-    transform=axes[2, 0].transAxes,
-)
+for row_label, ax_row in zip(["No Obs", "Obs", "Difference"], axes[:, 0]):
+    ax_row.text(
+        -0.07,
+        0.5,
+        row_label,
+        va="bottom",
+        ha="center",
+        rotation="vertical",
+        rotation_mode="anchor",
+        fontsize=12,
+        transform=ax_row.transAxes,
+    )
 
-# Add colour bars
-plt.colorbar(im0, ax=axes[0, -1], shrink=0.6, label=f"{variable} (m/s)")
-plt.colorbar(im1, ax=axes[1, -1], shrink=0.6, label=f"{variable} (m/s)")
-plt.colorbar(im2, ax=axes[2, -1], shrink=0.6, label=f"{variable} (m/s)")
+plt.colorbar(im0, ax=axes[0, -1], shrink=0.6, label=f"{plot_var} (m/s)")
+plt.colorbar(im1, ax=axes[1, -1], shrink=0.6, label=f"{plot_var} (m/s)")
+plt.colorbar(im2, ax=axes[2, -1], shrink=0.6, label=f"{plot_var} (m/s)")
 
 plt.tight_layout()
-plt.savefig("outputs/21_stormcast_sda_comparison.jpg", dpi=150)
+plt.savefig("outputs/03_stormcast_conus_sda_comparison.jpg", dpi=150)
 
 # %%
 # Ground Truth Comparison
 # -----------------------
-# Fetch HRRR analysis (ground truth) at each valid forecast time and compute
-# the absolute error of both the no-obs and obs forecasts. This shows whether
-# assimilation improves accuracy relative to the actual analysis.
+# Fetch HRRR analysis at each valid forecast time and compute the absolute error
+# of both the no-obs and assimilation forecasts. This shows whether assimilation
+# improves accuracy relative to the actual analysis.
 
 # %%
-truth = fetch_data(
+truth_x, truth_coords = fetch_data(
     hrrr,
     time=init_time,
-    variable=np.array(plot_vars),
+    variable=np.array([plot_var]),
     lead_time=np.array([np.timedelta64(h + 1, "h") for h in range(nsteps)]),
     device="cpu",
-    legacy=False,
 )
-ic["variable"] = np.array(plot_vars)
+truth_x, truth_coords = map_coords(
+    truth_x,
+    truth_coords,
+    {"hrrr_y": model.hrrr_y, "hrrr_x": model.hrrr_x},
+)
+# truth_x shape: (time=1, lead_time=nsteps, variable=1, H, W)
+truth_fields = truth_x[0, :, 0].numpy()  # (nsteps, H, W)
 
-truth = map_coords_xr(truth, {"hrrr_y": ic["hrrr_y"], "hrrr_x": ic["hrrr_x"]})
-truth_vals = truth.sel(variable=variable).values  # [nsteps, hrrr_y, hrrr_x]
-no_obs_vals = no_obs_ds["prediction"].sel(variable=variable).values
-obs_vals = obs_ds["prediction"].sel(variable=variable).values
-
-# Compute absolute errors against ground truth
-no_obs_err = np.abs(no_obs_vals[0] - truth_vals[0])
-obs_err = np.abs(obs_vals[0] - truth_vals[0])
+no_obs_err = np.abs(no_obs_fields - truth_fields)
+obs_err = np.abs(obs_fields - truth_fields)
 
 # %%
-# Plot absolute errors between the StormCast predictions and HRRR analysis ground truth.
-# In later time-steps it is clear that StormCast with SDA sampling using ISD station
-# observations has improved accuracy over the vanilla stormcast prediction.
+# Plot absolute errors between the StormCast-CONUS predictions and HRRR
+# analysis ground truth. In later time-steps the assimilated forecast typically
+# shows lower errors near the observation stations.
 
 # %%
 plt.close("all")
@@ -459,30 +443,20 @@ for step in range(nsteps):
         zorder=2,
     )
 
-axes[0, 0].text(
-    -0.07,
-    0.5,
-    "|No Obs - Truth|",
-    va="bottom",
-    ha="center",
-    rotation="vertical",
-    rotation_mode="anchor",
-    fontsize=12,
-    transform=axes[0, 0].transAxes,
-)
-axes[1, 0].text(
-    -0.07,
-    0.5,
-    "|Obs - Truth|",
-    va="bottom",
-    ha="center",
-    rotation="vertical",
-    rotation_mode="anchor",
-    fontsize=12,
-    transform=axes[1, 0].transAxes,
-)
+for row_label, ax_row in zip(["|No Obs − Truth|", "|Obs − Truth|"], axes[:, 0]):
+    ax_row.text(
+        -0.07,
+        0.5,
+        row_label,
+        va="bottom",
+        ha="center",
+        rotation="vertical",
+        rotation_mode="anchor",
+        fontsize=12,
+        transform=ax_row.transAxes,
+    )
 
-plt.colorbar(im0, ax=axes[0, -1], shrink=0.6, label=f"|Δ{variable}| (m/s)")
-plt.colorbar(im1, ax=axes[1, -1], shrink=0.6, label=f"|Δ{variable}| (m/s)")
+plt.colorbar(im0, ax=axes[0, -1], shrink=0.6, label=f"|Δ{plot_var}| (m/s)")
+plt.colorbar(im1, ax=axes[1, -1], shrink=0.6, label=f"|Δ{plot_var}| (m/s)")
 plt.tight_layout()
-plt.savefig("outputs/21_stormcast_sda_gt_comparison.jpg")
+plt.savefig("outputs/03_stormcast_conus_sda_gt_comparison.jpg", dpi=150)
